refactor: simplify Codex config and organize protocol tests
Remove unused Codex launch configuration, retain sidecar defaults, and move Responses/codec tests to the gateway layer. Document the configuration migration and preserve wrapper coverage.

Validation: 495 CPU tests passed; the single redis/redis failure also reproduces on clean base. Fresh rollout finished with reward/acc 1.0. Existing tool-call rejections and subsequent Transformer Engine PPO-update failure remain outside this cleanup.
